snap: bump to build-base: core24
Note the drastic change: we are migrating to grub as the bootloader!

In Noble, the boot0 binary for the Nezha board was split into one for
the LicheeRV and another for the Nezha. This complicates the story for
our gadget; we now have to ship two different versions of it. Testing
will reveal whether one gadget can still boot both boards, but for
now we will split some hairs. The introduction of this split has also
introduced some ugliness in how we determine which board to build for. I
am very much displeased by these changes.

In the short term, we will probably split the repository into a branch
for each board which both publish to the same snap under different track
names just to make local builds easier and cleaner, at the cost of a
slight amount of maintenance burden.

In the long term, we will investigate leveraging components to ship
separate bootloader files as separate assets.
Note that this idea is currently not officially spec'd or supported, so
this is a very long term goal.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>