u-boot/boot.scr.in: add template boot.scr.in
This file offers an example of a minimum-viable boot.scr file to be
executed by u-boot to boot Ubuntu Core.

The particulars of any of these environment variables or what kinds of
files are allowed to be loaded and executed are specific to whatever
u-boot binary you package with the gadget.

The final file you ship may barely resemble this file, and that is
fine. Simply attempt to enable an automated booting process which also
allows for potentially rolling back the kernel release or booting into
Recovery.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>