snap/snapcraft.yaml: add parts for u-boot
We definitely need to provide an example in the case we choose u-boot
as our bootloader in the gadget.yaml. Note that the u-boot part itself
should be expanded to include some general information on how to build
u-boot from source just in case a prebuilt binary is not available in
the archive.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>