snapcraft.yaml: Add a u-boot example
The u-boot example includes the case of either leveraging a package in
the archive or building u-boot from source. Because the latter case is
very target dependent, readers are instructed to refer to the official
board and u-boot documentation for how to accomplish this, as well as to
look at examples in other branches of this repository.

The comments have also been updated to include additional useful
information, and the tense has been modified slightly. Some small
formatting and typo errors were also resolved.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>