many: consolidate nezha and licheerv builds
As it turns out, the Sipeed LicheeRV and AllWinner Nezha can boot from
the same partition layout and u-boot binary. This simplifies everything
for this gadget greatly.

In the future it may be worth investigating why the u-boot binaries are
split by Canonical, but for now using a singular binary and layout works
to boot both devices.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>