secureboot: add example for implementing SecureBoot
SecureBoot and Full Disk Encryption are major value adds to Ubuntu Core,
but require potentially quite rigorous effort to enable. There isn't
robust support for OP-TEE across all hardware platforms, and as such the
target hardware must be selected carefully to ensure that SB and FDE are
feasible on the platform.

The example platform used here is the NXP i.MX8MM.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>